requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyterlab
pillow
opencv-python-headless
tqdm
openpyxl
ipykernel
transformers[torch]
torch
torchvision
xgboost